node-main/deps/v8/test/webkit/dfg-side-effect-assignment-osr-exit-expected.txt

# Copyright 2013 the V8 project authors. All rights reserved.
# Copyright (C) 2005, 2006, 2007, 2008, 2009 Apple Inc. All rights reserved.
#
# Redistribution and use in source and binary forms, with or without
# modification, are permitted provided that the following conditions
# are met:
# 1.  Redistributions of source code must retain the above copyright
#     notice, this list of conditions and the following disclaimer.
# 2.  Redistributions in binary form must reproduce the above copyright
#     notice, this list of conditions and the following disclaimer in the
#     documentation and/or other materials provided with the distribution.
#
# THIS SOFTWARE IS PROVIDED BY APPLE INC. AND ITS CONTRIBUTORS ``AS IS'' AND ANY
# EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
# WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
# DISCLAIMED. IN NO EVENT SHALL APPLE INC. OR ITS CONTRIBUTORS BE LIABLE FOR ANY
# DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
# (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
# LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
# ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
# (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
# SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Tests what happens if we OSR exit on an assignment that was part of a side-effecting bytecode instruction.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foo(bar) is 42
PASS count is 1
PASS foo(bar) is 42
PASS count is 2
PASS foo(bar) is 42
PASS count is 3
PASS foo(bar) is 42
PASS count is 4
PASS foo(bar) is 42
PASS count is 5
PASS foo(bar) is 42
PASS count is 6
PASS foo(bar) is 42
PASS count is 7
PASS foo(bar) is 42
PASS count is 8
PASS foo(bar) is 42
PASS count is 9
PASS foo(bar) is 42
PASS count is 10
PASS foo(bar) is 42
PASS count is 11
PASS foo(bar) is 42
PASS count is 12
PASS foo(bar) is 42
PASS count is 13
PASS foo(bar) is 42
PASS count is 14
PASS foo(bar) is 42
PASS count is 15
PASS foo(bar) is 42
PASS count is 16
PASS foo(bar) is 42
PASS count is 17
PASS foo(bar) is 42
PASS count is 18
PASS foo(bar) is 42
PASS count is 19
PASS foo(bar) is 42
PASS count is 20
PASS foo(bar) is 42
PASS count is 21
PASS foo(bar) is 42
PASS count is 22
PASS foo(bar) is 42
PASS count is 23
PASS foo(bar) is 42
PASS count is 24
PASS foo(bar) is 42
PASS count is 25
PASS foo(bar) is 42
PASS count is 26
PASS foo(bar) is 42
PASS count is 27
PASS foo(bar) is 42
PASS count is 28
PASS foo(bar) is 42
PASS count is 29
PASS foo(bar) is 42
PASS count is 30
PASS foo(bar) is 42
PASS count is 31
PASS foo(bar) is 42
PASS count is 32
PASS foo(bar) is 42
PASS count is 33
PASS foo(bar) is 42
PASS count is 34
PASS foo(bar) is 42
PASS count is 35
PASS foo(bar) is 42
PASS count is 36
PASS foo(bar) is 42
PASS count is 37
PASS foo(bar) is 42
PASS count is 38
PASS foo(bar) is 42
PASS count is 39
PASS foo(bar) is 42
PASS count is 40
PASS foo(bar) is 42
PASS count is 41
PASS foo(bar) is 42
PASS count is 42
PASS foo(bar) is 42
PASS count is 43
PASS foo(bar) is 42
PASS count is 44
PASS foo(bar) is 42
PASS count is 45
PASS foo(bar) is 42
PASS count is 46
PASS foo(bar) is 42
PASS count is 47
PASS foo(bar) is 42
PASS count is 48
PASS foo(bar) is 42
PASS count is 49
PASS foo(bar) is 42
PASS count is 50
PASS foo(bar) is 42
PASS count is 51
PASS foo(bar) is 42
PASS count is 52
PASS foo(bar) is 42
PASS count is 53
PASS foo(bar) is 42
PASS count is 54
PASS foo(bar) is 42
PASS count is 55
PASS foo(bar) is 42
PASS count is 56
PASS foo(bar) is 42
PASS count is 57
PASS foo(bar) is 42
PASS count is 58
PASS foo(bar) is 42
PASS count is 59
PASS foo(bar) is 42
PASS count is 60
PASS foo(bar) is 42
PASS count is 61
PASS foo(bar) is 42
PASS count is 62
PASS foo(bar) is 42
PASS count is 63
PASS foo(bar) is 42
PASS count is 64
PASS foo(bar) is 42
PASS count is 65
PASS foo(bar) is 42
PASS count is 66
PASS foo(bar) is 42
PASS count is 67
PASS foo(bar) is 42
PASS count is 68
PASS foo(bar) is 42
PASS count is 69
PASS foo(bar) is 42
PASS count is 70
PASS foo(bar) is 42
PASS count is 71
PASS foo(bar) is 42
PASS count is 72
PASS foo(bar) is 42
PASS count is 73
PASS foo(bar) is 42
PASS count is 74
PASS foo(bar) is 42
PASS count is 75
PASS foo(bar) is 42
PASS count is 76
PASS foo(bar) is 42
PASS count is 77
PASS foo(bar) is 42
PASS count is 78
PASS foo(bar) is 42
PASS count is 79
PASS foo(bar) is 42
PASS count is 80
PASS foo(bar) is 42
PASS count is 81
PASS foo(bar) is 42
PASS count is 82
PASS foo(bar) is 42
PASS count is 83
PASS foo(bar) is 42
PASS count is 84
PASS foo(bar) is 42
PASS count is 85
PASS foo(bar) is 42
PASS count is 86
PASS foo(bar) is 42
PASS count is 87
PASS foo(bar) is 42
PASS count is 88
PASS foo(bar) is 42
PASS count is 89
PASS foo(bar) is 42
PASS count is 90
PASS foo(bar) is 42
PASS count is 91
PASS foo(bar) is 42
PASS count is 92
PASS foo(bar) is 42
PASS count is 93
PASS foo(bar) is 42
PASS count is 94
PASS foo(bar) is 42
PASS count is 95
PASS foo(bar) is 42
PASS count is 96
PASS foo(bar) is 42
PASS count is 97
PASS foo(bar) is 42
PASS count is 98
PASS foo(bar) is 42
PASS count is 99
PASS foo(bar) is 42
PASS count is 100
PASS foo(bar) is 42
PASS count is 101
PASS foo(bar) is 42
PASS count is 102
PASS foo(bar) is 42
PASS count is 103
PASS foo(bar) is 42
PASS count is 104
PASS foo(bar) is 42
PASS count is 105
PASS foo(bar) is 42
PASS count is 106
PASS foo(bar) is 42
PASS count is 107
PASS foo(bar) is 42
PASS count is 108
PASS foo(bar) is 42
PASS count is 109
PASS foo(bar) is 42
PASS count is 110
PASS foo(bar) is 42
PASS count is 111
PASS foo(bar) is 42
PASS count is 112
PASS foo(bar) is 42
PASS count is 113
PASS foo(bar) is 42
PASS count is 114
PASS foo(bar) is 42
PASS count is 115
PASS foo(bar) is 42
PASS count is 116
PASS foo(bar) is 42
PASS count is 117
PASS foo(bar) is 42
PASS count is 118
PASS foo(bar) is 42
PASS count is 119
PASS foo(bar) is 42
PASS count is 120
PASS foo(bar) is 42
PASS count is 121
PASS foo(bar) is 42
PASS count is 122
PASS foo(bar) is 42
PASS count is 123
PASS foo(bar) is 42
PASS count is 124
PASS foo(bar) is 42
PASS count is 125
PASS foo(bar) is 42
PASS count is 126
PASS foo(bar) is 42
PASS count is 127
PASS foo(bar) is 42
PASS count is 128
PASS foo(bar) is 42
PASS count is 129
PASS foo(bar) is 42
PASS count is 130
PASS foo(bar) is 42
PASS count is 131
PASS foo(bar) is 42
PASS count is 132
PASS foo(bar) is 42
PASS count is 133
PASS foo(bar) is 42
PASS count is 134
PASS foo(bar) is 42
PASS count is 135
PASS foo(bar) is 42
PASS count is 136
PASS foo(bar) is 42
PASS count is 137
PASS foo(bar) is 42
PASS count is 138
PASS foo(bar) is 42
PASS count is 139
PASS foo(bar) is 42
PASS count is 140
PASS foo(bar) is 42
PASS count is 141
PASS foo(bar) is 42
PASS count is 142
PASS foo(bar) is 42
PASS count is 143
PASS foo(bar) is 42
PASS count is 144
PASS foo(bar) is 42
PASS count is 145
PASS foo(bar) is 42
PASS count is 146
PASS foo(bar) is 42
PASS count is 147
PASS foo(bar) is 42
PASS count is 148
PASS foo(bar) is 42
PASS count is 149
PASS foo(bar) is 42
PASS count is 150
PASS foo(bar) is 42
PASS count is 151
PASS foo(bar) is 42
PASS count is 152
PASS foo(bar) is 42
PASS count is 153
PASS foo(bar) is 42
PASS count is 154
PASS foo(bar) is 42
PASS count is 155
PASS foo(bar) is 42
PASS count is 156
PASS foo(bar) is 42
PASS count is 157
PASS foo(bar) is 42
PASS count is 158
PASS foo(bar) is 42
PASS count is 159
PASS foo(bar) is 42
PASS count is 160
PASS foo(bar) is 42
PASS count is 161
PASS foo(bar) is 42
PASS count is 162
PASS foo(bar) is 42
PASS count is 163
PASS foo(bar) is 42
PASS count is 164
PASS foo(bar) is 42
PASS count is 165
PASS foo(bar) is 42
PASS count is 166
PASS foo(bar) is 42
PASS count is 167
PASS foo(bar) is 42
PASS count is 168
PASS foo(bar) is 42
PASS count is 169
PASS foo(bar) is 42
PASS count is 170
PASS foo(bar) is 42
PASS count is 171
PASS foo(bar) is 42
PASS count is 172
PASS foo(bar) is 42
PASS count is 173
PASS foo(bar) is 42
PASS count is 174
PASS foo(bar) is 42
PASS count is 175
PASS foo(bar) is 42
PASS count is 176
PASS foo(bar) is 42
PASS count is 177
PASS foo(bar) is 42
PASS count is 178
PASS foo(bar) is 42
PASS count is 179
PASS foo(bar) is 42
PASS count is 180
PASS foo(bar) is 42
PASS count is 181
PASS foo(bar) is 42
PASS count is 182
PASS foo(bar) is 42
PASS count is 183
PASS foo(bar) is 42
PASS count is 184
PASS foo(bar) is 42
PASS count is 185
PASS foo(bar) is 42
PASS count is 186
PASS foo(bar) is 42
PASS count is 187
PASS foo(bar) is 42
PASS count is 188
PASS foo(bar) is 42
PASS count is 189
PASS foo(bar) is 42
PASS count is 190
PASS foo(bar) is 42
PASS count is 191
PASS foo(bar) is 42
PASS count is 192
PASS foo(bar) is 42
PASS count is 193
PASS foo(bar) is 42
PASS count is 194
PASS foo(bar) is 42
PASS count is 195
PASS foo(bar) is 42
PASS count is 196
PASS foo(bar) is 42
PASS count is 197
PASS foo(bar) is 42
PASS count is 198
PASS foo(bar) is 42
PASS count is 199
PASS foo(bar) is 42
PASS count is 200
PASS foo(bar) is 42
PASS count is 201
PASS foo(bar) is 42
PASS count is 202
PASS foo(bar) is 42
PASS count is 203
PASS foo(bar) is 42
PASS count is 204
PASS foo(bar) is 42
PASS count is 205
PASS foo(bar) is 42
PASS count is 206
PASS foo(bar) is 42
PASS count is 207
PASS foo(bar) is 42
PASS count is 208
PASS foo(bar) is 42
PASS count is 209
PASS foo(bar) is 42
PASS count is 210
PASS foo(bar) is 42
PASS count is 211
PASS foo(bar) is 42
PASS count is 212
PASS foo(bar) is 42
PASS count is 213
PASS foo(bar) is 42
PASS count is 214
PASS foo(bar) is 42
PASS count is 215
PASS foo(bar) is 42
PASS count is 216
PASS foo(bar) is 42
PASS count is 217
PASS foo(bar) is 42
PASS count is 218
PASS foo(bar) is 42
PASS count is 219
PASS foo(bar) is 42
PASS count is 220
PASS foo(bar) is 42
PASS count is 221
PASS foo(bar) is 42
PASS count is 222
PASS foo(bar) is 42
PASS count is 223
PASS foo(bar) is 42
PASS count is 224
PASS foo(bar) is 42
PASS count is 225
PASS foo(bar) is 42
PASS count is 226
PASS foo(bar) is 42
PASS count is 227
PASS foo(bar) is 42
PASS count is 228
PASS foo(bar) is 42
PASS count is 229
PASS foo(bar) is 42
PASS count is 230
PASS foo(bar) is 42
PASS count is 231
PASS foo(bar) is 42
PASS count is 232
PASS foo(bar) is 42
PASS count is 233
PASS foo(bar) is 42
PASS count is 234
PASS foo(bar) is 42
PASS count is 235
PASS foo(bar) is 42
PASS count is 236
PASS foo(bar) is 42
PASS count is 237
PASS foo(bar) is 42
PASS count is 238
PASS foo(bar) is 42
PASS count is 239
PASS foo(bar) is 42
PASS count is 240
PASS foo(bar) is 42
PASS count is 241
PASS foo(bar) is 42
PASS count is 242
PASS foo(bar) is 42
PASS count is 243
PASS foo(bar) is 42
PASS count is 244
PASS foo(bar) is 42
PASS count is 245
PASS foo(bar) is 42
PASS count is 246
PASS foo(bar) is 42
PASS count is 247
PASS foo(bar) is 42
PASS count is 248
PASS foo(bar) is 42
PASS count is 249
PASS foo(bar) is 42
PASS count is 250
PASS foo(bar) is 42
PASS count is 251
PASS foo(bar) is 42
PASS count is 252
PASS foo(bar) is 42
PASS count is 253
PASS foo(bar) is 42
PASS count is 254
PASS foo(bar) is 42
PASS count is 255
PASS foo(bar) is 42
PASS count is 256
PASS foo(bar) is 42
PASS count is 257
PASS foo(bar) is 42
PASS count is 258
PASS foo(bar) is 42
PASS count is 259
PASS foo(bar) is 42
PASS count is 260
PASS foo(bar) is 42
PASS count is 261
PASS foo(bar) is 42
PASS count is 262
PASS foo(bar) is 42
PASS count is 263
PASS foo(bar) is 42
PASS count is 264
PASS foo(bar) is 42
PASS count is 265
PASS foo(bar) is 42
PASS count is 266
PASS foo(bar) is 42
PASS count is 267
PASS foo(bar) is 42
PASS count is 268
PASS foo(bar) is 42
PASS count is 269
PASS foo(bar) is 42
PASS count is 270
PASS foo(bar) is 42
PASS count is 271
PASS foo(bar) is 42
PASS count is 272
PASS foo(bar) is 42
PASS count is 273
PASS foo(bar) is 42
PASS count is 274
PASS foo(bar) is 42
PASS count is 275
PASS foo(bar) is 42
PASS count is 276
PASS foo(bar) is 42
PASS count is 277
PASS foo(bar) is 42
PASS count is 278
PASS foo(bar) is 42
PASS count is 279
PASS foo(bar) is 42
PASS count is 280
PASS foo(bar) is 42
PASS count is 281
PASS foo(bar) is 42
PASS count is 282
PASS foo(bar) is 42
PASS count is 283
PASS foo(bar) is 42
PASS count is 284
PASS foo(bar) is 42
PASS count is 285
PASS foo(bar) is 42
PASS count is 286
PASS foo(bar) is 42
PASS count is 287
PASS foo(bar) is 42
PASS count is 288
PASS foo(bar) is 42
PASS count is 289
PASS foo(bar) is 42
PASS count is 290
PASS foo(bar) is 42
PASS count is 291
PASS foo(bar) is 42
PASS count is 292
PASS foo(bar) is 42
PASS count is 293
PASS foo(bar) is 42
PASS count is 294
PASS foo(bar) is 42
PASS count is 295
PASS foo(bar) is 42
PASS count is 296
PASS foo(bar) is 42
PASS count is 297
PASS foo(bar) is 42
PASS count is 298
PASS foo(bar) is 42
PASS count is 299
PASS foo(bar) is 42
PASS count is 300
PASS foo(bar) is 42
PASS count is 301
PASS foo(bar) is 42
PASS count is 302
PASS foo(bar) is 42
PASS count is 303
PASS foo(bar) is 42
PASS count is 304
PASS foo(bar) is 42
PASS count is 305
PASS foo(bar) is 42
PASS count is 306
PASS foo(bar) is 42
PASS count is 307
PASS foo(bar) is 42
PASS count is 308
PASS foo(bar) is 42
PASS count is 309
PASS foo(bar) is 42
PASS count is 310
PASS foo(bar) is 42
PASS count is 311
PASS foo(bar) is 42
PASS count is 312
PASS foo(bar) is 42
PASS count is 313
PASS foo(bar) is 42
PASS count is 314
PASS foo(bar) is 42
PASS count is 315
PASS foo(bar) is 42
PASS count is 316
PASS foo(bar) is 42
PASS count is 317
PASS foo(bar) is 42
PASS count is 318
PASS foo(bar) is 42
PASS count is 319
PASS foo(bar) is 42
PASS count is 320
PASS foo(bar) is 42
PASS count is 321
PASS foo(bar) is 42
PASS count is 322
PASS foo(bar) is 42
PASS count is 323
PASS foo(bar) is 42
PASS count is 324
PASS foo(bar) is 42
PASS count is 325
PASS foo(bar) is 42
PASS count is 326
PASS foo(bar) is 42
PASS count is 327
PASS foo(bar) is 42
PASS count is 328
PASS foo(bar) is 42
PASS count is 329
PASS foo(bar) is 42
PASS count is 330
PASS foo(bar) is 42
PASS count is 331
PASS foo(bar) is 42
PASS count is 332
PASS foo(bar) is 42
PASS count is 333
PASS foo(bar) is 42
PASS count is 334
PASS foo(bar) is 42
PASS count is 335
PASS foo(bar) is 42
PASS count is 336
PASS foo(bar) is 42
PASS count is 337
PASS foo(bar) is 42
PASS count is 338
PASS foo(bar) is 42
PASS count is 339
PASS foo(bar) is 42
PASS count is 340
PASS foo(bar) is 42
PASS count is 341
PASS foo(bar) is 42
PASS count is 342
PASS foo(bar) is 42
PASS count is 343
PASS foo(bar) is 42
PASS count is 344
PASS foo(bar) is 42
PASS count is 345
PASS foo(bar) is 42
PASS count is 346
PASS foo(bar) is 42
PASS count is 347
PASS foo(bar) is 42
PASS count is 348
PASS foo(bar) is 42
PASS count is 349
PASS foo(bar) is 42
PASS count is 350
PASS foo(bar) is 42
PASS count is 351
PASS foo(bar) is 42
PASS count is 352
PASS foo(bar) is 42
PASS count is 353
PASS foo(bar) is 42
PASS count is 354
PASS foo(bar) is 42
PASS count is 355
PASS foo(bar) is 42
PASS count is 356
PASS foo(bar) is 42
PASS count is 357
PASS foo(bar) is 42
PASS count is 358
PASS foo(bar) is 42
PASS count is 359
PASS foo(bar) is 42
PASS count is 360
PASS foo(bar) is 42
PASS count is 361
PASS foo(bar) is 42
PASS count is 362
PASS foo(bar) is 42
PASS count is 363
PASS foo(bar) is 42
PASS count is 364
PASS foo(bar) is 42
PASS count is 365
PASS foo(bar) is 42
PASS count is 366
PASS foo(bar) is 42
PASS count is 367
PASS foo(bar) is 42
PASS count is 368
PASS foo(bar) is 42
PASS count is 369
PASS foo(bar) is 42
PASS count is 370
PASS foo(bar) is 42
PASS count is 371
PASS foo(bar) is 42
PASS count is 372
PASS foo(bar) is 42
PASS count is 373
PASS foo(bar) is 42
PASS count is 374
PASS foo(bar) is 42
PASS count is 375
PASS foo(bar) is 42
PASS count is 376
PASS foo(bar) is 42
PASS count is 377
PASS foo(bar) is 42
PASS count is 378
PASS foo(bar) is 42
PASS count is 379
PASS foo(bar) is 42
PASS count is 380
PASS foo(bar) is 42
PASS count is 381
PASS foo(bar) is 42
PASS count is 382
PASS foo(bar) is 42
PASS count is 383
PASS foo(bar) is 42
PASS count is 384
PASS foo(bar) is 42
PASS count is 385
PASS foo(bar) is 42
PASS count is 386
PASS foo(bar) is 42
PASS count is 387
PASS foo(bar) is 42
PASS count is 388
PASS foo(bar) is 42
PASS count is 389
PASS foo(bar) is 42
PASS count is 390
PASS foo(bar) is 42
PASS count is 391
PASS foo(bar) is 42
PASS count is 392
PASS foo(bar) is 42
PASS count is 393
PASS foo(bar) is 42
PASS count is 394
PASS foo(bar) is 42
PASS count is 395
PASS foo(bar) is 42
PASS count is 396
PASS foo(bar) is 42
PASS count is 397
PASS foo(bar) is 42
PASS count is 398
PASS foo(bar) is 42
PASS count is 399
PASS foo(bar) is 42
PASS count is 400
PASS foo(bar) is 42
PASS count is 401
PASS foo(bar) is 42
PASS count is 402
PASS foo(bar) is 42
PASS count is 403
PASS foo(bar) is 42
PASS count is 404
PASS foo(bar) is 42
PASS count is 405
PASS foo(bar) is 42
PASS count is 406
PASS foo(bar) is 42
PASS count is 407
PASS foo(bar) is 42
PASS count is 408
PASS foo(bar) is 42
PASS count is 409
PASS foo(bar) is 42
PASS count is 410
PASS foo(bar) is 42
PASS count is 411
PASS foo(bar) is 42
PASS count is 412
PASS foo(bar) is 42
PASS count is 413
PASS foo(bar) is 42
PASS count is 414
PASS foo(bar) is 42
PASS count is 415
PASS foo(bar) is 42
PASS count is 416
PASS foo(bar) is 42
PASS count is 417
PASS foo(bar) is 42
PASS count is 418
PASS foo(bar) is 42
PASS count is 419
PASS foo(bar) is 42
PASS count is 420
PASS foo(bar) is 42
PASS count is 421
PASS foo(bar) is 42
PASS count is 422
PASS foo(bar) is 42
PASS count is 423
PASS foo(bar) is 42
PASS count is 424
PASS foo(bar) is 42
PASS count is 425
PASS foo(bar) is 42
PASS count is 426
PASS foo(bar) is 42
PASS count is 427
PASS foo(bar) is 42
PASS count is 428
PASS foo(bar) is 42
PASS count is 429
PASS foo(bar) is 42
PASS count is 430
PASS foo(bar) is 42
PASS count is 431
PASS foo(bar) is 42
PASS count is 432
PASS foo(bar) is 42
PASS count is 433
PASS foo(bar) is 42
PASS count is 434
PASS foo(bar) is 42
PASS count is 435
PASS foo(bar) is 42
PASS count is 436
PASS foo(bar) is 42
PASS count is 437
PASS foo(bar) is 42
PASS count is 438
PASS foo(bar) is 42
PASS count is 439
PASS foo(bar) is 42
PASS count is 440
PASS foo(bar) is 42
PASS count is 441
PASS foo(bar) is 42
PASS count is 442
PASS foo(bar) is 42
PASS count is 443
PASS foo(bar) is 42
PASS count is 444
PASS foo(bar) is 42
PASS count is 445
PASS foo(bar) is 42
PASS count is 446
PASS foo(bar) is 42
PASS count is 447
PASS foo(bar) is 42
PASS count is 448
PASS foo(bar) is 42
PASS count is 449
PASS foo(bar) is 42
PASS count is 450
PASS foo(bar) is "stuff"
PASS count is 451
PASS foo(bar) is "stuff"
PASS count is 452
PASS foo(bar) is "stuff"
PASS count is 453
PASS foo(bar) is "stuff"
PASS count is 454
PASS foo(bar) is "stuff"
PASS count is 455
PASS foo(bar) is "stuff"
PASS count is 456
PASS foo(bar) is "stuff"
PASS count is 457
PASS foo(bar) is "stuff"
PASS count is 458
PASS foo(bar) is "stuff"
PASS count is 459
PASS foo(bar) is "stuff"
PASS count is 460
PASS foo(bar) is "stuff"
PASS count is 461
PASS foo(bar) is "stuff"
PASS count is 462
PASS foo(bar) is "stuff"
PASS count is 463
PASS foo(bar) is "stuff"
PASS count is 464
PASS foo(bar) is "stuff"
PASS count is 465
PASS foo(bar) is "stuff"
PASS count is 466
PASS foo(bar) is "stuff"
PASS count is 467
PASS foo(bar) is "stuff"
PASS count is 468
PASS foo(bar) is "stuff"
PASS count is 469
PASS foo(bar) is "stuff"
PASS count is 470
PASS foo(bar) is "stuff"
PASS count is 471
PASS foo(bar) is "stuff"
PASS count is 472
PASS foo(bar) is "stuff"
PASS count is 473
PASS foo(bar) is "stuff"
PASS count is 474
PASS foo(bar) is "stuff"
PASS count is 475
PASS foo(bar) is "stuff"
PASS count is 476
PASS foo(bar) is "stuff"
PASS count is 477
PASS foo(bar) is "stuff"
PASS count is 478
PASS foo(bar) is "stuff"
PASS count is 479
PASS foo(bar) is "stuff"
PASS count is 480
PASS foo(bar) is "stuff"
PASS count is 481
PASS foo(bar) is "stuff"
PASS count is 482
PASS foo(bar) is "stuff"
PASS count is 483
PASS foo(bar) is "stuff"
PASS count is 484
PASS foo(bar) is "stuff"
PASS count is 485
PASS foo(bar) is "stuff"
PASS count is 486
PASS foo(bar) is "stuff"
PASS count is 487
PASS foo(bar) is "stuff"
PASS count is 488
PASS foo(bar) is "stuff"
PASS count is 489
PASS foo(bar) is "stuff"
PASS count is 490
PASS foo(bar) is "stuff"
PASS count is 491
PASS foo(bar) is "stuff"
PASS count is 492
PASS foo(bar) is "stuff"
PASS count is 493
PASS foo(bar) is "stuff"
PASS count is 494
PASS foo(bar) is "stuff"
PASS count is 495
PASS foo(bar) is "stuff"
PASS count is 496
PASS foo(bar) is "stuff"
PASS count is 497
PASS foo(bar) is "stuff"
PASS count is 498
PASS foo(bar) is "stuff"
PASS count is 499
PASS foo(bar) is "stuff"
PASS count is 500
PASS successfullyParsed is true

TEST COMPLETE
